requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
djangorestframework==3.16.1
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2

# Database Drivers (CRITICAL - must be installed for DB connection)
psycopg2-binary>=2.9.10
psycopg2>=2.9.10

# LLM / AI
ollama>=0.5.3
openai>=1.102.0

# Web / Frameworks
httpx>=0.28.1

# Core / Utilities
python-decouple>=3.8

# Optional database drivers (remove if not needed)
mssql-django>=1.6
PyMySQL>=1.1.2
pyodbc>=5.2.0
redis>=6.4.0

langgraph